requirements.txt
allauth-django==0.42.6
asgiref==3.2.10
astroid==2.4.2
autopep8==1.5.3
certifi==2020.6.20
chardet==3.0.4
defusedxml==0.6.0
Django==3.0.8
django-rest-auth==0.9.5
django-rest-registration==0.5.6
djangorestframework==3.11.0
idna==2.10
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysqlclient==2.0.1
oauthlib==3.1.0
pycodestyle==2.6.0
pylint==2.5.3
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typing==3.7.4.3
urllib3==1.25.9
wrapt==1.12.1
